Adjust calculator style
Changes in the Javascript file just amount to more fiddling with the
way chained operations are handled. Since I haven't figured the logic
out, I think it would be best to make a feature branch so that
I can play in a walled garden instead of making meaningless
changes. Real progress for this commit only happened in
the HTML and CSS files.